# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DatabaseType = typeof DatabaseType[keyof typeof DatabaseType];


export const DatabaseType = {
  postgres: 'postgres',
  mysql: 'mysql',
  sqlite: 'sqlite',
  sqlserver: 'sqlserver',
} as const;

export interface AccountUser {
  id: string;
  email: string;
  createdAt: string;
}

export interface SessionResponse {
  authenticated: boolean;
  user: AccountUser | null;
}

export interface RegisterInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 12
     * @maxLength 256
     */
  password: string;
}

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 256
     */
  password: string;
}

export interface DatabaseConnection {
  id: string;
  name: string;
  databaseType: DatabaseType;
  createdAt: string;
  /** @nullable */
  lastTestedAt: string | null;
}

export interface ConnectionInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  databaseType: DatabaseType;
  /**
     * @minLength 5
     * @maxLength 4096
     */
  connectionString: string;
}

export interface ConnectionTestInput {
  databaseType: DatabaseType;
  /**
     * @minLength 5
     * @maxLength 4096
     */
  connectionString: string;
}

export interface ConnectionTestResult {
  success: boolean;
  databaseType: DatabaseType;
  /** @minimum 0 */
  schemaCount: number;
  /** @minimum 0 */
  tableCount: number;
  message: string;
}

export interface ApiToken {
  id: string;
  connectionId: string;
  connectionName: string;
  label: string;
  tokenPrefix: string;
  createdAt: string;
  /** @nullable */
  lastUsedAt: string | null;
  /** @nullable */
  revokedAt: string | null;
}

export interface TokenInput {
  connectionId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
}

export interface TokenCreated {
  /** Plaintext bearer token; shown only once. */
  token: string;
  tokenInfo: ApiToken;
}

export interface DashboardSummary {
  user: AccountUser;
  /** @minimum 0 */
  connectionsCount: number;
  /** @minimum 0 */
  activeTokensCount: number;
  /** @minimum 0 */
  revokedTokensCount: number;
  recentConnections: DatabaseConnection[];
  recentTokens: ApiToken[];
}