# lib/api-zod/src/generated/types/apiToken.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiToken {
  id: string;
  connectionId: string;
  connectionName: string;
  label: string;
  tokenPrefix: string;
  createdAt: Date;
  /** @nullable */
  lastUsedAt: Date | null;
  /** @nullable */
  revokedAt: Date | null;
}