# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountUser } from './accountUser';
import type { ApiToken } from './apiToken';
import type { DatabaseConnection } from './databaseConnection';

export interface DashboardSummary {
  user: AccountUser;
  /** @minimum 0 */
  connectionsCount: number;
  /** @minimum 0 */
  activeTokensCount: number;
  /** @minimum 0 */
  revokedTokensCount: number;
  recentConnections: DatabaseConnection[];
  recentTokens: ApiToken[];
}