# lib/api-zod/src/generated/types/databaseConnection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */
import type { DatabaseType } from './databaseType';

export interface DatabaseConnection {
  id: string;
  name: string;
  databaseType: DatabaseType;
  createdAt: Date;
  /** @nullable */
  lastTestedAt: Date | null;
}