# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}