# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 12
     * @maxLength 256
     */
  password: string;
}