# lib/api-zod/src/generated/types/tokenInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Database Pilot account, connection, and MCP token API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenInput {
  connectionId: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
}